+++ socialstream/chat/__init__.py
@@ -1,4 +1,3 @@
-from .chat import chat_demo
 from .chat_omniscient import chat_demo as chat_demo_omniscient
 
-__all__ = ["chat_demo", "chat_demo_omniscient"]
+__all__ = ["chat_demo_omniscient"]
