+++ socialstream/chat/chat_omniscient.py
@@ -1,32 +1,25 @@
 import json
 import time
-from typing import cast
 
 import streamlit as st
-from sotopia.database import (
-    AgentProfile,
-    EnvAgentComboStorage,
-    EnvironmentProfile,
-    EpisodeLog,
-)
-from sotopia.envs import ParallelSotopiaEnv
-from sotopia.envs.parallel import (
-    render_text_for_agent,
-)
-from sotopia.messages import AgentAction
 
+from socialstream.chat.callbacks import (
+    agent_edit_callback_finegrained,
+    edit_callback,
+    other_choice_callback,
+)
+from socialstream.rendering_utils import (
+    compose_agent_messages,
+    compose_env_messages,
+    messageForRendering,
+    render_messages,
+)
 from socialstream.utils import (
     HUMAN_MODEL_NAME,
     MODEL_LIST,
     ActionState,
     EnvAgentProfileCombo,
-    _agent_profile_to_friendabove_self,
-    get_public_info,
-    get_secret_info,
     initialize_session_state,
-    messageForRendering,
-    print_current_speaker,
-    render_for_humans,
     set_from_env_agent_profile_combo,
     set_settings,
     step,
@@ -35,27 +28,6 @@ from socialstream.utils import (
 
 def chat_demo() -> None:
     initialize_session_state()
-
-    def other_choice_callback() -> None:
-        st.session_state.agent_models = [
-            st.session_state.agent1_model_choice,
-            st.session_state.agent2_model_choice,
-        ]
-        st.session_state.editable = st.session_state.edit_scenario
-        print("Editable: ", st.session_state.editable)
-        agent_choice_1 = st.session_state.agent_choice_1
-        agent_choice_2 = st.session_state.agent_choice_2
-        set_settings(
-            agent_choice_1=agent_choice_1,
-            agent_choice_2=agent_choice_2,
-            scenario_choice=st.session_state.scenario_choice,
-            user_agent_name="PLACEHOLDER",
-            agent_names=[
-                st.session_state.agent_choice_1,
-                st.session_state.agent_choice_2,
-            ],
-            reset_agents=False,
-        )
 
     def env_agent_choice_callback() -> None:
         if st.session_state.active:
@@ -156,105 +128,9 @@ def chat_demo() -> None:
                 key="agent2_model_choice",
             )
 
-    def edit_callback(key: str = "", reset_msgs: bool = False) -> None:
-        # set agent_goals and environment background
-        env_profiles: EnvironmentProfile = st.session_state.env.profile
-        scenario = env_profiles.scenario
-        agent_goals = env_profiles.agent_goals
-
-        match key:
-            case "edited_scenario":
-                scenario = st.session_state[key]
-            case "edited_goal_0":
-                agent_goals[0] = st.session_state[key]
-            case "edited_goal_1":
-                agent_goals[1] = st.session_state[key]
-            # case _:
-            #     raise ValueError(f"Invalid key: {key}")
-
-        env_profiles.scenario = scenario
-        env_profiles.agent_goals = agent_goals
-
-        print("Edited scenario: ", env_profiles.scenario)
-        print("Edited goals: ", env_profiles.agent_goals)
-
-        env_agent_combo = EnvAgentProfileCombo(
-            env=env_profiles,
-            agents=[agent.profile for agent in st.session_state.agents.values()],
-        )
-        set_from_env_agent_profile_combo(
-            env_agent_combo=env_agent_combo, reset_msgs=reset_msgs
-        )
-
-    def agent_edit_callback_finegrained(key: str = "", traits: list[str] = []) -> None:
-        agents = list(st.session_state.agents.values())
-        agents_info = [
-            {key: getattr(agent.profile, key) for key in traits} for agent in agents
-        ]
-        # "edited_agent_{agent_idx}_{trait_name}"
-        trait = key.split("-")[-1]
-        agent_idx = int(key.split("-")[-2])
-
-        agents_info[agent_idx][trait] = st.session_state[key]
-        agent_1_profile = AgentProfile(**agents_info[0])
-        agent_2_profile = AgentProfile(**agents_info[1])
-        print("Edited agent 1: ", agent_1_profile)
-        print("Edited agent 2: ", agent_2_profile)
-
-        env_agent_combo = EnvAgentProfileCombo(
-            env=st.session_state.env.profile,
-            agents=[agent_1_profile, agent_2_profile],
-        )
-        set_from_env_agent_profile_combo(
-            env_agent_combo=env_agent_combo, reset_msgs=False
-        )
-
-    def agent_edit_callback(key: str = "") -> None:
-        agents = list(st.session_state.agents.values())
-        agents_info = [
-            {
-                "first_name": agent.profile.first_name,
-                "last_name": agent.profile.last_name,
-                "public_info": get_public_info(agent.profile, display_name=False),
-                "secret": get_secret_info(agent.profile, display_name=False),
-            }
-            for agent in agents
-        ]
-
-        match key:
-            case "edited_agent_0":
-                agents_info[0]["public_info"] = st.session_state[key]
-            case "edited_agent_1":
-                agents_info[1]["public_info"] = st.session_state[key]
-            case "edited_secret_0":
-                agents_info[0]["secret"] = st.session_state[key]
-            case "edited_secret_1":
-                agents_info[1]["secret"] = st.session_state[key]
-
-        agent_1_profile = AgentProfile(**agents_info[0])
-        agent_2_profile = AgentProfile(**agents_info[1])
-        print("Edited agent 1: ", agent_1_profile)
-        print("Edited agent 2: ", agent_2_profile)
-
-        env_agent_combo = EnvAgentProfileCombo(
-            env=st.session_state.env.profile,
-            agents=[agent_1_profile, agent_2_profile],
-        )
-        set_from_env_agent_profile_combo(
-            env_agent_combo=env_agent_combo, reset_msgs=False
-        )
-
     with st.expander("Check your social task!", expanded=True):
-        agent_infos = compose_agent_messages()
-        env_info, goals_info = compose_env_messages()
-        agent_public_infos = [
-            get_public_info(agent.profile, display_name=False)
-            for agent in st.session_state.agents.values()
-        ]
-        agent_secret_infos = [
-            get_secret_info(agent.profile, display_name=False)
-            for agent in st.session_state.agents.values()
-        ]
+        agent_infos = compose_agent_messages(agents=st.session_state.agents)
+        env_info, goals_info = compose_env_messages(env=st.session_state.env)
 
         if st.session_state.editable:
             st.text_area(
@@ -452,7 +328,13 @@ def chat_demo() -> None:
         with st.spinner("Evaluating..."):
             step()
 
-    messages = render_messages()
+    messages = render_messages(
+        env=st.session_state.env,
+        agent_list=list(st.session_state.agents.values()),
+        messages=st.session_state.messages,
+        reasoning=st.session_state.reasoning,
+        rewards=st.session_state.rewards,
+    )
     tag_for_eval = ["Agent 1", "Agent 2", "General"]
     chat_history = [
         message for message in messages if message["role"] not in tag_for_eval
@@ -531,48 +413,3 @@ def streamlit_rendering(messages: list[messageForRendering]) -> None:
                 st.markdown(content.replace("\n", "<br />"), unsafe_allow_html=True)
             else:
                 st.markdown(content.replace("\n", "<br />"), unsafe_allow_html=True)
-
-
-def compose_env_messages() -> tuple[str, list[str]]:
-    env: ParallelSotopiaEnv = st.session_state.env
-    env_profile = env.profile
-    env_to_render = env_profile.scenario
-    goals_to_render = env_profile.agent_goals
-
-    return env_to_render, goals_to_render
-
-
-def compose_agent_messages() -> list[str]:  # type: ignore
-    agents = st.session_state.agents
-
-    agent_to_render = [
-        render_text_for_agent(
-            raw_text=_agent_profile_to_friendabove_self(
-                agent.profile, agent_id, display_name=False
-            ),
-            agent_id=st.session_state.human_agent_idx,
-        )
-        for agent_id, agent in enumerate(agents.values())
-    ]
-    return agent_to_render
-
-
-def render_messages() -> list[messageForRendering]:
-    env = st.session_state.env
-    agent_list = list(st.session_state.agents.values())
-
-    epilog = EpisodeLog(
-        environment=env.profile.pk,
-        agents=[agent.profile.pk for agent in agent_list],
-        tag="tmp",
-        models=[env.model_name, agent_list[0].model_name, agent_list[1].model_name],
-        messages=[
-            [(m[0], m[1], m[2].to_natural_language()) for m in messages_in_turn]
-            for messages_in_turn in st.session_state.messages
-        ],
-        reasoning=st.session_state.reasoning,
-        rewards=st.session_state.rewards,
-        rewards_prompt="",
-    )
-    rendered_messages = render_for_humans(epilog)
-    return rendered_messages
