+++ socialstream/utils.py
@@ -24,13 +24,6 @@ from sotopia.envs.parallel import (
 )
 from sotopia.messages import AgentAction, Observation
 
-
-class messageForRendering(TypedDict):
-    role: str
-    type: str
-    content: str
-
-
 HUMAN_MODEL_NAME = "human"
 MODEL_LIST = [
     "gpt-4o-mini",
@@ -87,137 +80,6 @@ class EnvAgentProfileCombo:
     def __init__(self, env: EnvironmentProfile, agents: list[AgentProfile]) -> None:
         self.env = env
         self.agents = agents
-
-
-def parse_reasoning(reasoning: str, num_agents: int) -> tuple[list[str], str]:
-    """Parse the reasoning string into a dictionary."""
-    sep_token = "SEPSEP"
-    for i in range(1, num_agents + 1):
-        reasoning = (
-            reasoning.replace(f"Agent {i} comments:\n", sep_token)
-            .strip(" ")
-            .strip("\n")
-        )
-    all_chunks = reasoning.split(sep_token)
-    general_comment = all_chunks[0].strip(" ").strip("\n")
-    comment_chunks = all_chunks[-num_agents:]
-
-    return comment_chunks, general_comment
-
-
-def render_for_humans(episode: EpisodeLog) -> list[messageForRendering]:
-    """Generate a list of messages for human-readable version of the episode log."""
-
-    messages_for_rendering: list[messageForRendering] = []
-
-    for idx, turn in enumerate(episode.messages):
-        is_observation_printed = False
-
-        if idx == 0:
-            assert (
-                len(turn) >= 2
-            ), "The first turn should have at least environment messages"
-
-            messages_for_rendering.append(
-                {"role": "Background Info", "type": "info", "content": turn[0][2]}
-            )
-            messages_for_rendering.append(
-                {"role": "Background Info", "type": "info", "content": turn[1][2]}
-            )
-            messages_for_rendering.append(
-                {"role": "System", "type": "divider", "content": "Start Simulation"}
-            )
-
-        for sender, receiver, message in turn:
-            if not is_observation_printed and "Observation:" in message and idx != 0:
-                extract_observation = message.split("Observation:")[1].strip()
-                if extract_observation:
-                    messages_for_rendering.append(
-                        {
-                            "role": "Observation",
-                            "type": "observation",
-                            "content": extract_observation,
-                        }
-                    )
-                is_observation_printed = True
-
-            if receiver == "Environment":
-                if sender != "Environment":
-                    if "did nothing" in message:
-                        continue
-                    elif "left the conversation" in message:
-                        messages_for_rendering.append(
-                            {
-                                "role": "Environment",
-                                "type": "leave",
-                                "content": f"{sender} left the conversation",
-                            }
-                        )
-                    else:
-                        if "said:" in message:
-                            message = message.split("said:")[1].strip()
-                            messages_for_rendering.append(
-                                {"role": sender, "type": "said", "content": message}
-                            )
-                        else:
-                            message = message.replace("[action]", "")
-                            messages_for_rendering.append(
-                                {"role": sender, "type": "action", "content": message}
-                            )
-                else:
-                    messages_for_rendering.append(
-                        {
-                            "role": "Environment",
-                            "type": "environment",
-                            "content": message,
-                        }
-                    )
-
-    messages_for_rendering.append(
-        {"role": "System", "type": "divider", "content": "End Simulation"}
-    )
-
-    reasoning_per_agent, general_comment = parse_reasoning(
-        episode.reasoning,
-        len(
-            set(
-                msg["role"]
-                for msg in messages_for_rendering
-                if msg["type"] in {"said", "action"}
-            )
-        ),
-    )
-
-    if general_comment == "":
-        return messages_for_rendering[:-1]
-
-    messages_for_rendering.append(
-        {"role": "General", "type": "comment", "content": general_comment}
-    )
-
-    for idx, reasoning in enumerate(reasoning_per_agent):
-        reasoning_lines = reasoning.split("\n")
-        new_reasoning = ""
-        for reasoning_line in reasoning_lines:
-            dimension = reasoning_line.split(":")[0]
-            new_reasoning += (
-                (f"**{dimension}**: {':'.join(reasoning_line.split(':')[1:])}" + "\n")
-                if dimension != ""
-                else reasoning_line + "\n"
-            )
-        messages_for_rendering.append(
-            {
-                "role": f"Agent {idx + 1}",
-                "type": "comment",
-                "content": f"**Agent {idx + 1} reasoning**:\n{new_reasoning}\n\n**Rewards**: {str(episode.rewards[idx])}",
-            }
-        )
-
-    for item in messages_for_rendering:
-        item["content"] = item["content"].replace("$", "\\$")
-    # print("Replaced content", messages_for_rendering)
-
-    return messages_for_rendering
 
 
 def async_to_sync(async_func: callable) -> callable:
@@ -433,8 +295,6 @@ def step(user_input: str | None = None) -> None:
 
         actions.append(action)
 
-    actions = cast(list[AgentAction], actions)
-
     for idx, agent_name in enumerate(st.session_state.env.agents):
         agent_messages[agent_name] = actions[idx]
         st.session_state.messages[-1].append(
@@ -492,56 +352,5 @@ def step(user_input: str | None = None) -> None:
     done = all(terminated.values())
 
 
-def _map_gender_to_adj(gender: str) -> str:
-    gender_to_adj = {
-        "Man": "male",
-        "Woman": "female",
-        "Nonbinary": "nonbinary",
-    }
-    if gender:
-        return gender_to_adj[gender]
-    else:
-        return ""
-
-
 # def _agent_profile_to_friendabove_self(profile: AgentProfile, agent_id: int) -> str:
 #     return f"{profile.first_name} {profile.last_name} is a {profile.age}-year-old {_map_gender_to_adj(profile.gender)} {profile.occupation.lower()}. {profile.gender_pronoun} pronouns. {profile.public_info} Personality and values description: {profile.personality_and_values} <p viewer='agent_{agent_id}'>{profile.first_name}'s secrets: {profile.secret}. Big five type: {profile.big_five}</p>"
-
-
-def agent_profile_to_public_info(
-    profile: AgentProfile, display_name: bool = False
-) -> str:
-    base_str: str = f"a {profile.age}-year-old {_map_gender_to_adj(profile.gender)} {profile.occupation.lower()}. {profile.gender_pronoun} pronouns. {profile.public_info} Personality and values description: {profile.personality_and_values}."
-    if display_name:
-        return f"{profile.first_name} {profile.last_name} is {base_str}"
-    else:
-        return base_str.capitalize()
-
-
-def agent_profile_to_secret_info(
-    profile: AgentProfile, display_name: bool = False
-) -> str:
-    if display_name:
-        return f"{profile.first_name}'s secrets: {profile.secret}."
-    else:
-        return f"Secrets: {profile.secret}."
-
-
-def get_public_info(profile: AgentProfile, display_name: bool = True) -> str:
-    if profile.age == 0:
-        return profile.public_info
-    else:
-        return agent_profile_to_public_info(profile, display_name=display_name)
-
-
-def get_secret_info(profile: AgentProfile, display_name: bool = True) -> str:
-    if profile.age == 0:
-        return profile.secret
-    else:
-        return agent_profile_to_secret_info(profile, display_name=display_name)
-
-
-def _agent_profile_to_friendabove_self(
-    profile: AgentProfile, agent_id: int, display_name: bool = True
-) -> str:
-    return f"{get_public_info(profile=profile, display_name=display_name)}<p viewer='agent_{agent_id}'>{get_secret_info(profile=profile, display_name=display_name)}</p>"
